# Full traversal in dask.order