--- a/dask/tests/test_distributed.py
+++ b/dask/tests/test_distributed.py
@@ -3,6 +3,7 @@ distributed = pytest.importorskip('distributed')
 
 from functools import partial
 import inspect
+from operator import add
 from tornado import gen
 
 import dask
@@ -111,3 +112,17 @@ def test_serializable_groupby_agg(c, s, a, b):
     result = ddf.groupby('y').agg('count')
 
     yield c.compute(result)
+
+
+def test_futures_in_graph(loop):
+    with cluster() as (s, [a, b]):
+        with Client(s['address'], loop=loop) as c:
+            x, y = delayed(1), delayed(2)
+            xx = delayed(add)(x, x)
+            yy = delayed(add)(y, y)
+            xxyy = delayed(add)(xx, yy)
+
+            xxyy2 = c.persist(xxyy)
+            xxyy3 = delayed(add)(xxyy2, 10)
+
+            assert xxyy3.compute(get=c.get) == ((1 + 1) + (2 + 2)) + 10