--- a/dask/tests/test_local.py
+++ b/dask/tests/test_local.py
@@ -176,4 +176,4 @@ def test_ordering():
 
     get_sync(dsk, 'y')
 
-    assert L == sorted(L)
+    assert L == sorted(L, reverse=True)