--- a/dask/tests/test_order.py
+++ b/dask/tests/test_order.py
@@ -1,5 +1,6 @@
 import pytest
 
+import dask
 from dask.order import ndependencies, order
 from dask.core import get_deps
 from dask.utils_test import add, inc
@@ -244,7 +245,7 @@ def test_gh_3055():
     dsk = dict(w.__dask_graph__())
     o = order(dsk)
     L = [o[k] for k in w.__dask_keys__()]
-    assert sorted(L) == L[::-1]
+    assert sorted(L) == L[::-1] or sorted(L) == L
 
 
 def test_type_comparisions_ok(abcde):
@@ -271,3 +272,85 @@ def test_prefer_short_dependents(abcde):
     o = order(dsk)
     assert o[d] < o[b]
     assert o[e] < o[b]
+
+
+def test_run_smaller_sections(abcde):
+    """
+            aa
+           / |
+      b   d  bb dd
+     / \ /|  | /
+    a   c e  cc
+
+    Prefer to run acb first because then we can get that out of the way
+    """
+    a, b, c, d, e = abcde
+    aa, bb, cc, dd = [x * 2 for x in [a, b, c, d]]
+
+    expected = [a, c, b, e, d, cc, bb, aa, dd]
+
+    log = []
+    def f(x):
+        def _(*args):
+            log.append(x)
+        return _
+
+    dsk = {a: (f(a),),
+           c: (f(c),),
+           e: (f(e),),
+           cc: (f(cc),),
+           b: (f(b), a, c),
+           d: (f(d), c, e),
+           bb: (f(bb), cc),
+           aa: (f(aa), d, bb),
+           dd: (f(dd), cc)}
+
+    dask.get(dsk, [aa, b, dd])  # trigger computation
+
+    assert log == expected
+
+
+def test_local_parents_of_reduction(abcde):
+    """
+
+            c1
+            |
+        b1  c2
+        |  /|
+    a1  b2  c3
+    |  /|
+    a2  b3
+    |
+    a3
+
+    Prefer to finish a1 stack before proceding to b2
+    """
+    a, b, c, d, e = abcde
+    a1, a2, a3 = [a + i for i in '123']
+    b1, b2, b3 = [b + i for i in '123']
+    c1, c2, c3 = [c + i for i in '123']
+
+    expected = [a3, a2, a1,
+                b3, b2, b1,
+                c3, c2, c1]
+
+    log = []
+    def f(x):
+        def _(*args):
+            log.append(x)
+        return _
+
+    dsk = {a3: (f(a3),),
+           a2: (f(a2), a3),
+           a1: (f(a1), a2),
+           b3: (f(b3),),
+           b2: (f(b2), b3, a2),
+           b1: (f(b1), b2),
+           c3: (f(c3),),
+           c2: (f(c2), c3, b2),
+           c1: (f(c1), c2)}
+
+    o = order(dsk)
+    dask.get(dsk, [a1, b1, c1])  # trigger computation
+
+    assert log == expected