--- a/dask/utils_test.py
+++ b/dask/utils_test.py
@@ -1,5 +1,7 @@
 from __future__ import absolute_import, division, print_function
 
+import pytest
+
 
 def inc(x):
     return x + 1
@@ -107,10 +109,14 @@ class GetFunctionTestMixin(object):
         assert self.get(d, 'z') == 4
 
     def test_get_stack_limit(self):
-        d = dict(('x%d' % (i + 1), (inc, 'x%d' % i)) for i in range(10000))
+        d = {'x%d' % (i + 1): (inc, 'x%d' % i) for i in range(10000)}
         d['x0'] = 0
         assert self.get(d, 'x10000') == 10000
-        # introduce cycle
+
+    @pytest.mark.skip(reason="We are no longer robust to graph cycles")
+    def test_robust_to_cycles(self):
+        d = {'x%d' % (i + 1): (inc, 'x%d' % i) for i in range(10000)}
+        d['x0'] = 0
         d['x5000'] = (inc, 'x5001')
 
         try: