--- a/dask/utils_test.py
+++ b/dask/utils_test.py
@@ -1,7 +1,5 @@
 from __future__ import absolute_import, division, print_function
 
-import pytest
-
 
 def inc(x):
     return x + 1
@@ -112,25 +110,6 @@ class GetFunctionTestMixin(object):
         d = {'x%d' % (i + 1): (inc, 'x%d' % i) for i in range(10000)}
         d['x0'] = 0
         assert self.get(d, 'x10000') == 10000
-
-    @pytest.mark.skip(reason="We are no longer robust to graph cycles")
-    def test_robust_to_cycles(self):
-        d = {'x%d' % (i + 1): (inc, 'x%d' % i) for i in range(10000)}
-        d['x0'] = 0
-        d['x5000'] = (inc, 'x5001')
-
-        try:
-            self.get(d, 'x10000')
-        except (RuntimeError, ValueError) as e:
-            if isinstance(e, RuntimeError):
-                assert str(e) == 'Cycle detected in Dask: x5001->x5000->x5001'
-            elif isinstance(e, ValueError):
-                assert str(e).startswith('Found no accessible jobs in dask')
-        else:
-            msg = 'dask with infinite cycle should have raised an exception.'
-            assert False, msg
-
-        assert self.get(d, 'x4999') == 4999
 
     def test_with_sharedict(self):
         from .sharedict import ShareDict