@mrocklin 2018-01-15
So this test seems to conflict with the following situation:

```python
    def test_avoid_upwards_branching(abcde):
        """
             a1
             |
             a2
             |
             a3    d1
            /  \  /
          b1    c1
          |     |
          b2    c2
                |
                c3
    
        Prefer b1 over c1 because it won't stick around waiting for d1 to complete
        """
        a, b, c, d, e = abcde
        dsk = {(a, 1): (f, (a, 2)),
               (a, 2): (f, (a, 3)),
               (a, 3): (f, (b, 1), (c, 1)),
               (b, 1): (f, (b, 2)),
               (c, 1): (f, (c, 2)),
               (c, 2): (f, (c, 3)),
               (d, 1): (f, (c, 1))}
    
        o = order(dsk)    
```

In this test we want to select `b`, but doing so requires that we start descending from `a`.  The test in this PR asks that we start from short top-nodes because they can be finished more quickly, allowing their children to be released once a long-running co-parent chain starts running.  The test in this PR asks us to start descending from short top-nodes.  The test in this comment asks that we start from long top-nodes.

This brings up the question of not doing top-down depth first, but something that wanders around a bit more or swaps values during traversal.