@mrocklin 2018-01-16
OK, there is a solution here that solves a number of problems modestly well.  It is sacrifices perfection in a few cases for "good enough" in more.  The traversal is more complex now and may be more expensive.  I'll have to do some benchmarking.  Here are the visualizations for a couple of troublesome graphs.

![dask](https://user-images.githubusercontent.com/306380/35012108-86c964d6-fad6-11e7-8c53-a820c5094877.png)
![gh3055](https://user-images.githubusercontent.com/306380/35012109-86df75fa-fad6-11e7-9fa8-a43a697a4a17.png)

```python
import matplotlib.pyplot as plt

import dask.array as da
n = 10
x  = da.random.normal(size=(n, 100), chunks=(1, 100))
y  = da.random.normal(size=(n,), chunks=(1,))
xy = (x * y[:, None]).cumsum(axis=0)
xx = (x[:, None, :] * x[:, :, None]).cumsum(axis=0)
beta = da.stack([da.linalg.solve(xx[i], xy[i]) for i in range(xx.shape[0])],
                axis=0)
ey = (x * beta).sum(axis=1)
ey.visualize('dask.png', color='order', node_attr=dict(penwidth='8'),
             optimize_graph=True, cmap=plt.cm.RdBu)


A, B = 10, 99
x = da.random.normal(size=(A, B), chunks=(1, None))
for _ in range(2):
    y = (x[:, None, :] * x[:, :, None]).cumsum(axis=0)
    x = x.cumsum(axis=0)
w = (y * x[:, None]).sum(axis=(1,2))

w.visualize('gh3055.png', color='order', node_attr=dict(penwidth='8'),
            cmap=plt.cm.RdBu)
```