@mrocklin 2018-01-16
@eriknw and @jcrist might both find this PR of interest