@mrocklin 2018-01-16
Anecdotal timing information

```python
%%time
n = 100
x  = da.random.normal(size=(n, 100), chunks=(1, 100))
y  = da.random.normal(size=(n,), chunks=(1,))
xy = (x * y[:, None]).cumsum(axis=0)
xx = (x[:, None, :] * x[:, :, None]).cumsum(axis=0)
beta = da.stack([da.linalg.solve(xx[i], xy[i]) for i in range(xx.shape[0])],
                axis=0)
ey = (x * beta).sum(axis=1)
CPU times: user 239 ms, sys: 4.21 ms, total: 243 ms
Wall time: 233 ms
```

```python
%time dsk = dict(ey.__dask_graph__())
CPU times: user 74.1 ms, sys: 0 ns, total: 74.1 ms
Wall time: 73.2 ms
```

### Master: 25ms

```python
%timeit _ = order(dsk)
10 loops, best of 3: 25.9 ms per loop
```

### This branch: 39ms

```python
%timeit _ = order(dsk)
10 loops, best of 3: 39.3 ms per loop
```

This is non-trivial, but possibly still worth it.