@mrocklin 2018-02-01
Merging this soon if there are no further comments.  Alternatively @jcrist, you're probably the most likely to review this.  Have you had a chance to look it over?