@mrocklin 2018-02-01
@jcrist ok to merge on passed tests?